requirements.txt
SpeechRecognition
pyaudio
openai
python-dotenv
cohere
Flask
flask-cors
langdetect
huggingface
transformers
torch

django
django-rest-framework

cchardet
chardet